seminars/seminar_14/requirements.txt
python-docx==1.2.0
transformers
sentence-transformers
# torch==2.6.0+cu126
# torchvision==0.21.0
qdrant-client
openai